refactor(sketching): fold sketchLib + sketchUtils into sketch + sketchFns
Layer 3 audit F8 — `sketching/` had four files with overlapping names
(`sketchLib.ts`, `sketchUtils.ts`, `sketchFns.ts`, plus `sketch.ts` and
`sketches.ts`) where two were trivial:

- `sketchLib.ts` (52 LOC) — just declared `SketchInterface`
- `sketchUtils.ts` (23 LOC) — just two `wrapSketchData*` factories

Move the interface to `sketch.ts` (where the implementing class lives)
and the wrap helpers to `sketchFns.ts` (the functional layer). Public
`SketchInterface` re-exports updated in `src/index.ts` and
`src/sketching.ts` to point at the new location.

Net: 2 fewer files, no duplicated names, behaviour unchanged.
