--- src/sketching/sketchUtils.ts
@@ -1,23 +1,0 @@
-/**
- * Shared utilities for sketch creation and manipulation.
- */
-
-import type { PointInput } from '@/core/types.js';
-import type { SketchData } from '@/2d/blueprints/lib.js';
-import Sketch from './sketch.js';
-import CompoundSketch from './compoundSketch.js';
-
-/** Wrap SketchData into a Sketch instance. */
-export function wrapSketchData(data: SketchData): Sketch {
-  const opts: { defaultOrigin?: PointInput; defaultDirection?: PointInput } = {};
-  if (data.defaultOrigin) opts.defaultOrigin = data.defaultOrigin;
-  if (data.defaultDirection) opts.defaultDirection = data.defaultDirection;
-  const sketch = new Sketch(data.wire, opts);
-  if (data.baseFace) sketch.baseFace = data.baseFace;
-  return sketch;
-}
-
-/** Wrap an array of SketchData into a CompoundSketch. */
-export function wrapSketchDataArray(dataArr: SketchData[]): CompoundSketch {
-  return new CompoundSketch(dataArr.map(wrapSketchData));
-}
